Änderungen 17/6

+++ skelett/src/GLMatrix.cpp
@@ -26,7 +26,7 @@ void GLMatrix::setColumn(int i, const GLVector &rhs) {
   }
 }
 
-void GLMatrix::setValue(const int row, const int column, const float value) {
+void GLMatrix::setValue(const int row, const int column, const double value) {
   mMatrix[row * 4 + column] = value;
 }
